+++ agent_circuit_breaker/core/__init__.py
@@ -1,5 +1,6 @@
 """Core async pipeline primitives."""
 
+from .benchmarks import PipelineBenchmark, benchmark_pipeline
 from .context import AgentContext, current_context, context_scope
 from .pipeline import PipelineEngine
 from .results import GuardResult, PipelineResult
@@ -9,8 +10,10 @@ __all__ = [
     "AgentCircuitBreaker",
     "AgentContext",
     "GuardResult",
+    "PipelineBenchmark",
     "PipelineEngine",
     "PipelineResult",
+    "benchmark_pipeline",
     "context_scope",
     "current_context",
 ]